requirements.txt
doc2txt
gradio
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
tqdm